Move the sleeping bit to the top, fix the wake test on 32-bit
The sleeping flag now takes the top bit of wake_requested, apart from the
request bits. The wake stress test counted polls in a 64-bit atomic, which
32-bit targets don't have; a u32 is plenty.

+++ src/ev/test/wake.zig
@@ -43,7 +43,7 @@ test "Loop: wakes from another thread always end the poll they race with" {
         loop: Loop = undefined,
         ready: std.atomic.Value(bool) = .init(false),
         stop: std.atomic.Value(bool) = .init(false),
-        polls: std.atomic.Value(u64) = .init(0),
+        polls: std.atomic.Value(u32) = .init(0),
         wake_done: std.atomic.Value(bool) = .init(false),
         keep_alive: Timer = .init(.{ .duration = .zero }),
